Add files via upload
GUIlivard is a simple tool to Manage all the elements of a Game, a Graphical Interface, or anything else on a NumWorks Calculator (with Python).

Here is the content of this first commit:
— Image Compression: The image can be compressed and uncompressed on a single string to be sending on the calculator.
— Sprite Manager: It manages all sprites. We can ask it for the collision(s) between two of them or with the cursor.

Now, these are features that I want to add on next updates:
— The Screen Class: It will be able to draw the sprite on it and manage screen movements.
— The Event Manager: the programmer could ask the key pressed or any other user interaction he wants but also set a bunch of recursive events.
— Update the Working Test file for all features we add.


Here is the main line I want the project to follow:
— GUIlivard must HELP the developers to make their ideas.
— GUIlivard could be available on all calculators with python on it.
— We must follow the notices and advices of the user to do the correct updates.


Finally, please remember that this commit is the first. The project can change many times. The architecture, the functionalities and all the rest is not definitive!